nsxiv-rifle: more correct fallback handling
originally we checked the extension to make sure the argument is an
image to avoid edge cases like this:

	./a        # is a dir
	./a.jpg

where running `nsxiv-rifle a` would open up `a.jpg` instead of `a` the
directory.

instead of doing extension check in two places, now we just have a
is_img_extension() function which checks for extension
case-insensitively via grep.

in short:

* reduce some code duplication
* nsxiv will print an error msg for invalid files instead of the script
  silently exiting

this is probably a tad bit slower, but IMO worth the trade.